Move bitset-related code to a new crate `bitset`
