+++ select/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [dependencies]
 base-file = { path = "../base-file" }
+bitset = { path = "../bitset" }
 lock = { path = "../dependencies/kernel-sync" }
 syscall = { path = "../syscall" }
 task-trampoline = { path = "../task-trampoline" }
